pyproject.toml
# FcstLabPro

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "fcstlabpro"
version = "0.1.0"
description = "Bitcoin price reversal prediction experiment platform"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "lightgbm>=4.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pyyaml>=6.0",
    "joblib>=1.3",
    "tabulate>=0.9",
    "requests>=2.31",
    "yfinance>=0.2",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
